fix:allow to only owner and admin to delete comment

+++ src/main/java/ac/kr/smu/lookie/socialworker/controller/CommentController.java
@@ -2,11 +2,13 @@ package ac.kr.smu.lookie.socialworker.controller;
 
 import ac.kr.smu.lookie.socialworker.domain.Comment;
 import ac.kr.smu.lookie.socialworker.domain.Post;
+import ac.kr.smu.lookie.socialworker.domain.User;
 import ac.kr.smu.lookie.socialworker.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
@@ -36,7 +38,7 @@ public class CommentController {
     public ResponseEntity<?> postRecomment(@RequestBody Comment comment, @PathVariable("postId") Long postId,
                                            @PathVariable("preCommentId") Long preCommentId) {
         comment.setPost(Post.builder().id(postId).build());
-        
+
         EntityModel<Comment> body = EntityModel.of(commentService.register(preCommentId, comment));
         body.add(linkTo(methodOn(CommentController.class).postRecomment(comment, postId, preCommentId)).withSelfRel());
 
@@ -44,7 +46,10 @@ public class CommentController {
     }
 
     @DeleteMapping("/{commentId}")
-    public ResponseEntity deleteComment(@PathVariable("commentId") Long commentId) {
+    public ResponseEntity deleteComment(@PathVariable("commentId") Long commentId, @AuthenticationPrincipal User user) {
+        if(!user.equals(commentService.getComment(commentId).getUser()) || !user.getRoles().contains("ADMIN"))
+            return ResponseEntity.status(403).build();
+        
         return ResponseEntity.ok(commentService.delete(commentId));
     }
 }
